% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trial_geom.R
\name{calc_angle}
\alias{calc_angle}
\title{Function to get the heading angle between two points:}
\usage{
calc_angle(x, y)
}
\arguments{
\item{x}{point feature of class 'sf'}

\item{y}{point feature of class 'sf'}
}
\description{
This take in any two point features of class 'sf' and returns
 heading angle between the two points:
}
\examples{
\dontrun{

points <- sf::read_sf("data/Points.gpkg")
x <- sf::st_coordinates(points[1, 4])
y <- sf::st_coordinates(points[2, 4])
}
calc_angle(x, y)
}
\keyword{Angle,}
\keyword{Features,}
\keyword{Simple}
\keyword{heading,}
\keyword{sf}
