% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trial_geom.R
\name{calc_direction}
\alias{calc_direction}
\title{Function to get the heading angle between all consecutive points:}
\usage{
calc_direction(df)
}
\arguments{
\item{df}{data frame of class 'sf'}
}
\description{
This function take a point data frame of class 'sf' and return the heading angle between the
consecutive points. In other words, the result is a list of the direction of travel for each point:
}
\examples{
library(sf)
points <- read_sf("data/Points.gpkg")
calc_direction(points)
\dontrun{

}
}
\keyword{Angle,}
\keyword{Features,}
\keyword{Heading,}
\keyword{Simple}
\keyword{sf}
