% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trial_geom.R
\name{calc_distance}
\alias{calc_distance}
\title{Function to get distances between consecutive points:}
\usage{
calc_distance(sf_obj, max = FALSE)
}
\arguments{
\item{sf_obj}{object of class 'sf'}

\item{max}{conditional variable - default = FALSE}
}
\description{
This goes through every point in sf_obj and sums the distances between each consecutive point. If parameter max is FALSE sf_obj should
be a point feature. If max = TRUE, sf_obj should be a polygon and it will return the diagonal distance of the field.
}
\examples{
#Default
points =  sf::read_sf('./data/Points.gpkg')
DIFMR::calc_distance(points)
#max = TRUE
field = DIFMR::fields[4,]
DIFMR::calc_distance(field, max = TRUE)

}
\keyword{Distance,}
\keyword{Features,}
\keyword{Simple}
\keyword{sf}
