% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trial_geom.R
\name{calc_max_dist}
\alias{calc_max_dist}
\title{Function to get the diagonal distance of the bounding box of sf objects:
This takes in an sf object and calculates the distance between oppsoite corners of the bounding box}
\usage{
calc_max_dist(sf_obj)
}
\arguments{
\item{sf_obj}{of class 'sf' representing a rectangle}
}
\description{
Function to get the diagonal distance of the bounding box of sf objects:
This takes in an sf object and calculates the distance between oppsoite corners of the bounding box
}
\examples{
\dontrun{

}
}
\keyword{Distance,}
\keyword{Features,}
\keyword{Simple}
\keyword{sf}
