% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/par.R
\name{calculate_bootstrap_ci}
\alias{calculate_bootstrap_ci}
\title{This function calculates the population attributable risk's (PAR) credibility
interval using a Bootstrap
The function samples from the Multinomial distribution to estimate the
posterior distribution.}
\usage{
calculate_bootstrap_ci(x, interval = 0.95, sample_count = 10000)
}
\arguments{
\item{x}{A vector containing the values of a, b, c, and d
in this order. Where
a: The count of rows where both exposure and outcome are 1.
b: The count of rows where exposure is 1 and outcome is 0.
c: The count of rows where exposure is 0 and outcome is 1.
d: The count of rows where both exposure and outcome are 0.}

\item{interval}{Interval for the confidence interval (default is 0.95).
Possible values are between 0 and 1.}

\item{sample_count}{Number of samples to draw from the Dirichlet
distribution.}
}
\value{
Matrix containing the lower and upper bounds of the credibility
interval. The first column contains the lower bound and the second column
contains the upper bound.
}
\description{
This function calculates the population attributable risk's (PAR) credibility
interval using a Bootstrap
The function samples from the Multinomial distribution to estimate the
posterior distribution.
}
\examples{
# Example usage:
x <- extract_abcd(data, "exposure", "outcome")
calculate_bootstrap_ci(x, 0.99, 5000)

}
