% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/par.R
\name{calculate_par}
\alias{calculate_par}
\title{This function calculates the population attributable risk (PAR)
for a binary exposure and outcome.
PAR = P(D+) - P(D+| E-) = (a + c)/(a + b + c + d) - (c + d)/c}
\usage{
calculate_par(x)
}
\arguments{
\item{x}{A vector containing the values of a, b, c, and d
in this order. Where
a: The count of rows where both exposure and outcome are 1.
b: The count of rows where exposure is 1 and outcome is 0.
c: The count of rows where exposure is 0 and outcome is 1.
d: The count of rows where both exposure and outcome are 0.}
}
\value{
The population attributable risk (PAR).
}
\description{
This function calculates the population attributable risk (PAR)
for a binary exposure and outcome.
PAR = P(D+) - P(D+| E-) = (a + c)/(a + b + c + d) - (c + d)/c
}
\examples{
# Example usage:
x <- c(1, 2, 3, 4)
PAR <- calculate_par(x)
}
