% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/par.R
\name{compile_all}
\alias{compile_all}
\title{Compile all functions
This function compiles all functions in the package. This is useful for
optimizing the performance of the functions.}
\usage{
compile_all()
}
\description{
Compile all functions
This function compiles all functions in the package. This is useful for
optimizing the performance of the functions.
}
