% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trial_geom.R
\name{creat_lines}
\alias{creat_lines}
\title{Function to create trials around points:}
\usage{
creat_lines(la, lb, np, center = FALSE, invp = TRUE)
}
\description{
Function to create trials around points:
}
\keyword{Design}
\keyword{Plots,}
\keyword{Trial}
