% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trial_geom.R
\name{creat_polygons}
\alias{creat_polygons}
\title{Function to generate polygon from each point:
TODO: Split in simpler functions}
\usage{
creat_polygons(pts, w)
}
\arguments{
\item{pts}{of class 'sf' to intersect}

\item{w}{of class 'sf' to be intersected}
}
\description{
This take in any two sets of sf geometries and returns
 the intersects in the same way as the sp over:
}
\examples{
\dontrun{

}
}
\keyword{Features,}
\keyword{Intersect,}
\keyword{Over,}
\keyword{Simple}
\keyword{sf}
