% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmp2.R
\name{create_points_hv}
\alias{create_points_hv}
\title{Function to average to lines:}
\usage{
create_points_hv(l, nrows, ncols)
}
\arguments{
\item{obj}{feature of class 'sf'}
}
\description{
This take in any features of class 'sf' and returns
 the rotate version of the geometries:
}
\keyword{Affine,}
\keyword{Angle,}
\keyword{Features,}
\keyword{Rotation,}
\keyword{Simple}
\keyword{sf}
