% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trial_geom.R
\name{expand_lines}
\alias{expand_lines}
\title{Function to expand both extremities of a line:}
\usage{
expand_lines(lines, dst = 10000)
}
\arguments{
\item{line}{of class 'sfc' to be expanded}

\item{length}{of class 'numeric' to expand line by}
}
\description{
Function to expand both extremities of a line:
}
\examples{
\dontrun{
line <- ablines[1, ]
line_exp <- expand_line(line, 1000)
st_length(line)
st_length(line_exp)
}
}
\keyword{Features,}
\keyword{Intersect,}
\keyword{Over,}
\keyword{Simple}
\keyword{sf}
