% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/par.R
\name{extract_abcd}
\alias{extract_abcd}
\title{This function extracts values from the specified exposure and outcome
columns in the provided dataset. This is done by counting the number of
each possible combination of exposure and outcome}
\usage{
extract_abcd(data, exposure_col, outcome_col)
}
\arguments{
\item{data}{Dataset.}

\item{exposure_col}{String, name of the exposure column in the dataset.}

\item{outcome_col}{String , name of the outcome column in the dataset.}
}
\value{
List containing:
x_1e1d: The count of rows where both exposure and outcome are 1.
x_1e0d: The count of rows where exposure is 1 and outcome is 0.
x_0e1d: The count of rows where exposure is 0 and outcome is 1.
x_0e0d: The count of rows where both exposure and outcome are 0.
}
\description{
This function extracts values from the specified exposure and outcome
columns in the provided dataset. This is done by counting the number of
each possible combination of exposure and outcome
}
\examples{
# Example usage:
data <- data.frame(exposure = c(0, 0, 1, 1, 0, 1, 0, 1),
                   outcome = c(0, 1, 0, 1, 0, 1, 1, 0))
result <- extract_abcd(data, "exposure", "outcome")

}
