% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trial_geom.R
\name{get_coverage}
\alias{get_coverage}
\title{Function to generate polygon coverage over a field:}
\usage{
get_coverage(path_lines, width, field)
}
\arguments{
\item{path_lines}{of class 'sf' containing parallel lines drawn across field}

\item{width}{of class 'numeric' representing distance between lines}

\item{field}{of class 'sf'}
}
\description{
This take in any two sets of sf geometries and returns
 the intersects in the same way as the sp over:
}
\examples{
\dontrun{

}
}
\keyword{Features,}
\keyword{Intersect,}
\keyword{Over,}
\keyword{Simple}
\keyword{sf}
