% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trial_geom.R
\name{get_subpols}
\alias{get_subpols}
\title{Function to create the subpolygons in each pass:}
\usage{
get_subpols(path_lines, path_pols, plot_width, subplot_length)
}
\arguments{
\item{path_lines}{of class 'sf' containing parallel lines}

\item{path_pols}{of class 'sf' containing polygons along the lines}
}
\description{
This takes in the path lines and path polygons
and divides the polygons into subpolygons
}
\examples{
\dontrun{

}
}
\keyword{Features,}
\keyword{Intersect,}
\keyword{Over,}
\keyword{Simple}
\keyword{sf}
