% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trial_geom.R
\name{group_subpols}
\alias{group_subpols}
\title{Function to create the experimental units based on the subpolygons:}
\usage{
group_subpols(spols, plot_length, subplot_length, max_gap = Inf,
  min_length = 0, fixed_length = FALSE, invert_side = FALSE)
}
\arguments{
\item{spols}{of class 'sf'}

\item{plot_length}{of class 'numeric'}

\item{subplot_length}{of class 'numeric'}

\item{min_length}{of class 'numeric'}
}
\description{
This take in any two sets of sf geometries and returns
 the intersects in the same way as the sp over:
}
\examples{
\dontrun{

}
}
\keyword{Features,}
\keyword{Simple}
\keyword{sf}
