% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trial_geom.R
\name{make_headland}
\alias{make_headland}
\title{Function to make the field headland.}
\usage{
make_headland(field, width1, width2 = 0, angle = pi/2, inv = FALSE)
}
\arguments{
\item{field}{the field boundary of class 'sf'}

\item{pols}{the trial design polygons of class 'sf'}

\item{width}{the size of the headland}
}
\description{
This function creates a constant buffer around the field boundary.
FOr more complex types of headland, it is suggested to use graphical tools
such as QGIS
}
\examples{
plot(make_headland(fields[4, ], 10, 30))
plot(make_headland(fields[4, ], 10, 30, pi))
}
\keyword{Border}
\keyword{Buffer,}
\keyword{Design,}
\keyword{Headland,}
\keyword{Trial}
