% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trial_geom.R
\name{make_line}
\alias{make_line}
\title{Function to create a line based in a starting point and
an angle or two points:}
\usage{
make_line(p1, p2 = NA, angle = 0, p4s = NA)
}
\arguments{
\item{p1}{of class 'matrix' as point 1}

\item{p2}{of class 'matrix' as point 2}

\item{angle}{of class 'numeric' to be calculated with p1}

\item{p4s}{of class 'crs'}
}
\description{
This take in any two sets of sf geometries and returns
 the intersects in the same way as the sp over:
}
\examples{
\dontrun{

}
}
\keyword{Features,}
\keyword{Intersect,}
\keyword{Over,}
\keyword{Simple}
\keyword{sf}
