% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trial_geom.R
\name{make_trial}
\alias{make_trial}
\title{Function to create trials around points:}
\usage{
make_trial(pols)
}
\description{
Function to create trials around points:
}
\keyword{Design}
\keyword{Plots,}
\keyword{Trial}
