% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trial_geom.R
\name{parallel_lines}
\alias{parallel_lines}
\title{Function to generate parallel straight lines:}
\usage{
parallel_lines(line, width, field, offset_path = 0, max_dist = NA,
  clip = FALSE)
}
\arguments{
\item{line}{of class 'sfc' to draw lines parallel to}

\item{width}{of class 'numeric' to separate parallel lines by}

\item{field}{of class 'sf'}

\item{max_dist}{of class 'numeric'}
}
\description{
This takes in a line and draws parallel lines to it
across the field, separated by the distance given in width
}
\examples{
\dontrun{

}
}
\keyword{Features,}
\keyword{Intersect,}
\keyword{Over,}
\keyword{Simple}
\keyword{sf}
