% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trial_geom.R
\name{select_plots}
\alias{select_plots}
\title{Function to create trials around points:}
\usage{
select_plots(field, path_lines, points, width, length, path_width = NA,
  subplot_length = NA)
}
\arguments{
\item{field}{the field boundary of class 'sf'}

\item{path_lines}{'sf' data frame containing parallel lines drawn across field}

\item{points}{'sf' data frame containing points}

\item{width}{the dimension of the field perpendiclar to \code{path_lines}}

\item{length}{the dimension of the field parallel to \code{path_lines}}

\item{path_width}{Optional. Width calculated in units of \code{path_width} if specified; otherwise, in meters}

\item{subplot_length}{Optional. Length calculated in units of \code{subplot_length} if specified; otherwise, in meters}
}
\description{
This function creates a trial geometry with given dimensions
around each point in \code{points}.
}
\examples{
data(points)
select_plots(field, path_lines, points, 100, 200)
}
\keyword{Design}
\keyword{Plots,}
\keyword{Trial}
