% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trial_geom.R
\name{sfc_as_cols}
\alias{sfc_as_cols}
\title{Function to rotate a polygon by a given angle:}
\usage{
sfc_as_cols(x, xy_names = c("x", "y"))
}
\arguments{
\item{obj}{feature of class 'sf'}

\item{a}{angle (radians)}
}
\description{
This take in any features of class 'sf' and returns
 the rotate version of the geometries:
}
\examples{

sfc_as_cols(fields[1, ])
}
\keyword{Affine,}
\keyword{Angle,}
\keyword{Features,}
\keyword{Rotation,}
\keyword{Simple}
\keyword{sf}
