% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmp2.R
\name{st_average_lines}
\alias{st_average_lines}
\title{Function to average to lines in batch:}
\usage{
st_average_lines(ptsa, ptsb, ii, label)
}
\arguments{
\item{obj}{feature of class 'sf'}
}
\description{
This take in any features of class 'sf' and returns
 the rotate version of the geometries:
}
\keyword{Affine,}
\keyword{Angle,}
\keyword{Features,}
\keyword{Rotation,}
\keyword{Simple}
\keyword{sf}
