% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trial_geom.R
\name{st_over}
\alias{st_over}
\title{Complements the over function in sf:}
\usage{
st_over(x, y)
}
\arguments{
\item{x}{of class 'sf' to intersect}

\item{y}{of class 'sf' to be intersected}
}
\description{
This take in any two sets of sf geometries and returns
 the intersects in the same way as the sp over:
}
\examples{
\dontrun{
st_over()
}
}
\keyword{Features,}
\keyword{Intersect,}
\keyword{Over,}
\keyword{Simple}
\keyword{sf}
