% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmp2.R
\name{st_rev_line}
\alias{st_rev_line}
\title{Function to rotate a polygon by a given angle:}
\usage{
st_rev_line(obj)
}
\arguments{
\item{obj}{feature of class 'sf'}
}
\description{
This take in any features of class 'sf' and returns
 the rotate version of the geometries:
}
\keyword{Affine,}
\keyword{Angle,}
\keyword{Features,}
\keyword{Rotation,}
\keyword{Simple}
\keyword{sf}
