% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trial_geom.R
\name{st_rotate}
\alias{st_rotate}
\title{Function to rotate a polygon by a given angle:}
\usage{
st_rotate(obj, a)
}
\arguments{
\item{obj}{feature of class 'sf'}

\item{a}{angle (radians)}
}
\description{
This take in any features of class 'sf' and returns
 the rotate version of the geometries:
}
\examples{

st_rotate(fields[1, ], pi / 2)
}
\keyword{Affine,}
\keyword{Angle,}
\keyword{Features,}
\keyword{Rotation,}
\keyword{Simple}
\keyword{sf}
