% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trial_geom.R
\name{st_utm}
\alias{st_utm}
\title{Tranform any projection to UTM for a sf object:}
\usage{
st_utm(sf_obj)
}
\arguments{
\item{sf_obj}{of class 'sf'}
}
\description{
This take in any sf geometries and returns
 a new sf geometry with the right UTM zone:
}
\examples{
\dontrun{
st_utm(fields)
}
}
\keyword{Features,}
\keyword{Projection,}
\keyword{Simple}
\keyword{UTM,}
\keyword{sf}
