% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmp2.R
\name{to_borders}
\alias{to_borders}
\title{Function to average to lines:}
\usage{
to_borders(pol, nv = 4)
}
\arguments{
\item{obj}{feature of class 'sf'}
}
\description{
This take in any features of class 'sf' and returns
 the rotate version of the geometries:
}
\keyword{Affine,}
\keyword{Angle,}
\keyword{Features,}
\keyword{Rotation,}
\keyword{Simple}
\keyword{sf}
