% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trial_geom.R
\name{to_line}
\alias{to_line}
\title{Function to rotate a polygon by a given angle:}
\usage{
to_line(pol)
}
\arguments{
\item{obj}{feature of class 'sf'}

\item{a}{angle (radians)}
}
\description{
This take in any features of class 'sf' and returns
 the rotate version of the geometries:
}
\examples{

to_line(fields[1, ])
}
\keyword{Affine,}
\keyword{Angle,}
\keyword{Features,}
\keyword{Rotation,}
\keyword{Simple}
\keyword{sf}
